+++ src/primitives/cylinder.js
@@ -7,6 +7,7 @@ import {
 	LineSegments,
 	Mesh,
 	ShaderMaterial,
+	RawShaderMaterial,
 	UniformsLib
 } from '../../vendors/three.js';
 
@@ -47,15 +48,98 @@ export default function ({ color, coords, edgeForm = {}, opacity = 1, radius = 1
 				opacity: { value: opacity }
 			},
 			vertexShader: `
-				attribute vec3 cylinderBegin;
-				attribute vec3 cylinderEnd;
+				in vec3 cylinderBegin;
+				in vec3 cylinderEnd;
 
-				varying vec3 vLightFront;
-				varying vec3 vIndirectFront;
+				uniform vec3 ambientLightColor;
+				uniform vec3 diffuse;
+				uniform float opacity;
 
-				#include <common>
-				#include <bsdfs>
-				#include <lights_pars_begin>
+				out vec4 vColor;
+
+				#define RECIPROCAL_PI 0.3183098861837907
+				#define saturate(a) clamp(a, 0.0, 1.0)
+
+				struct IncidentLight {
+					vec3 color;
+					vec3 direction;
+				};
+
+				struct GeometricContext {
+					vec3 position;
+					vec3 normal;
+				};
+
+				float getDistanceAttenuation(const in float lightDistance, const in float cutoffDistance, const in float decayExponent) {
+					if (cutoffDistance > 0.0 && decayExponent > 0.0) {
+						return pow(saturate(- lightDistance / cutoffDistance + 1.0), decayExponent);
+					}
+					return 1.0;
+				}
+
+				#if NUM_DIR_LIGHTS > 0
+					struct DirectionalLight {
+						vec3 direction;
+						vec3 color;
+					};
+
+					uniform DirectionalLight directionalLights[NUM_DIR_LIGHTS];
+
+					void getDirectionalLightInfo(const in DirectionalLight directionalLight, out IncidentLight light) {
+						light.color = directionalLight.color;
+						light.direction = directionalLight.direction;
+					}
+				#endif
+				#if NUM_POINT_LIGHTS > 0
+					struct PointLight {
+						vec3 position;
+						vec3 color;
+						float distance;
+						float decay;
+					};
+
+					uniform PointLight pointLights[NUM_POINT_LIGHTS];
+
+					void getPointLightInfo(const in PointLight pointLight, const in GeometricContext geometry, out IncidentLight light) {
+						vec3 lVector = pointLight.position - geometry.position;
+
+						light.direction = normalize(lVector);
+						light.color = pointLight.color * getDistanceAttenuation(length(lVector), pointLight.distance, pointLight.decay);
+					}
+				#endif
+				#if NUM_SPOT_LIGHTS > 0
+					struct SpotLight {
+						vec3 position;
+						vec3 direction;
+						vec3 color;
+						float distance;
+						float decay;
+						float coneCos;
+						float penumbraCos;
+					};
+
+					uniform SpotLight spotLights[NUM_SPOT_LIGHTS];
+
+					float getSpotAttenuation(const in float coneCosine, const in float penumbraCosine, const in float angleCosine) {
+						return smoothstep(coneCosine, penumbraCosine, angleCosine);
+					}
+
+					void getSpotLightInfo(const in SpotLight spotLight, const in GeometricContext geometry, out IncidentLight light) {
+						vec3 lVector = spotLight.position - geometry.position;
+
+						light.direction = normalize(lVector);
+
+						float angleCos = dot(light.direction, spotLight.direction);
+
+						float spotAttenuation = getSpotAttenuation(spotLight.coneCos, spotLight.penumbraCos, angleCos);
+
+						if (spotAttenuation > 0.0) {
+							light.color = spotLight.color * spotAttenuation * getDistanceAttenuation(length(lVector), spotLight.distance, spotLight.decay);
+						} else {
+							light.color = vec3(0.0);
+						}
+					}
+				#endif
 
 				void main() {
 					vec3 z = normalize(cylinderBegin - cylinderEnd);
@@ -79,26 +163,45 @@ export default function ({ color, coords, edgeForm = {}, opacity = 1, radius = 1
 
 					gl_Position = projectionMatrix * mvPosition;
 
-					vec3 transformedNormal = normalMatrix * normal;
+					GeometricContext geometry = GeometricContext(
+						mvPosition.xyz,
+						normalize(normalMatrix * normal)
+					);
 
-					#include <lights_lambert_vertex>
+					vec3 light = ambientLightColor;
+
+					IncidentLight directLight;
+
+					#if NUM_POINT_LIGHTS > 0
+						for (int i = 0; i < NUM_POINT_LIGHTS; i++) {
+							getPointLightInfo(pointLights[i], geometry, directLight);
+
+							light += saturate(dot(geometry.normal, directLight.direction)) * directLight.color;
+						}
+					#endif
+					#if NUM_SPOT_LIGHTS > 0
+						for (int i = 0; i < NUM_SPOT_LIGHTS; i++) {
+							getSpotLightInfo(spotLights[i], geometry, directLight);
+
+							light += saturate(dot(geometry.normal, directLight.direction)) * directLight.color;
+						}
+					#endif
+					#if NUM_DIR_LIGHTS > 0
+						for (int i = 0; i < NUM_DIR_LIGHTS; i++) {
+							getDirectionalLightInfo(directionalLights[i], directLight);
+
+							light += saturate(dot(geometry.normal, directLight.direction)) * directLight.color;
+						}
+					#endif
+
+					vColor = vec4(light * diffuse * RECIPROCAL_PI, opacity);
 				}
 			`,
 			fragmentShader: `
-				uniform vec3 diffuse;
-				uniform float opacity;
-
-				varying vec3 vLightFront;
-				varying vec3 vIndirectFront;
-
-				#include <common>
-				#include <bsdfs>
+				in lowp vec4 vColor;
 
 				void main() {
-					gl_FragColor = vec4(
-						(vLightFront + vIndirectFront) * BRDF_Lambert(diffuse),
-						opacity
-					);
+					pc_fragColor = vColor;
 				}
 			`
 		})
@@ -294,13 +397,17 @@ export default function ({ color, coords, edgeForm = {}, opacity = 1, radius = 1
 
 	const edges = new LineSegments(
 		edgesGeometry,
-		new ShaderMaterial({
+		new RawShaderMaterial({
 			uniforms: {
 				color: { value: edgeForm.color ?? [0, 0, 0] }
 			},
-			vertexShader: `
-				attribute vec3 cylinderBegin;
-				attribute vec3 cylinderEnd;
+			vertexShader: `#version 300 es
+				in vec3 position;
+				in vec3 cylinderBegin;
+				in vec3 cylinderEnd;
+
+				uniform mat4 projectionMatrix;
+				uniform mat4 modelViewMatrix;
 
 				void main() {
 					vec3 z = normalize(cylinderBegin - cylinderEnd);
@@ -323,11 +430,13 @@ export default function ({ color, coords, edgeForm = {}, opacity = 1, radius = 1
 					gl_Position = projectionMatrix * modelViewMatrix * cylinderMatrix * vec4(position, 1);
 				}
 			`,
-			fragmentShader: `
-				uniform vec3 color;
+			fragmentShader: `#version 300 es
+				uniform lowp vec3 color;
+
+				out lowp vec4 pc_fragColor;
 
 				void main() {
-					gl_FragColor = vec4(color, 1);
+					pc_fragColor = vec4(color, 1.0);
 				}
 			`
 		})