+++ src/primitives/polygon.js
@@ -4,6 +4,7 @@ import {
 	DoubleSide,
 	Mesh,
 	Quaternion,
+	RawShaderMaterial,
 	ShaderMaterial,
 	Shape,
 	ShapeGeometry,
@@ -164,7 +165,7 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 				opacity: { value: opacity }
 			},
 			vertexShader: `
-				varying vec3 vViewPosition;
+				out vec3 vViewPosition;
 
 				void main() {
 					vec4 mvPosition = modelViewMatrix * vec4(position, 1.0);
@@ -175,11 +176,12 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 				}
 			`,
 			fragmentShader: `
+				in vec3 vViewPosition;
+				in vec3 vNormal;
+
 				uniform vec3 diffuse;
 				uniform float opacity;
 				uniform vec3 ambientLightColor;
-
-				varying vec3 vViewPosition;
 
 				#define RECIPROCAL_PI 0.3183098861837907
 				#define saturate(a) clamp(a, 0.0, 1.0)
@@ -187,16 +189,6 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 				struct IncidentLight {
 					vec3 color;
 					vec3 direction;
-				};
-
-				struct ReflectedLight {
-					vec3 directDiffuse;
-					vec3 indirectDiffuse;
-				};
-
-				struct GeometricContext {
-					vec3 position;
-					vec3 normal;
 				};
 
 				float getDistanceAttenuation(const in float lightDistance, const in float cutoffDistance, const in float decayExponent) {
@@ -267,16 +259,17 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 					}
 				#endif
 
-				void RE_Direct(const in IncidentLight directLight, const in GeometricContext geometry, const in vec3 diffuseColor, inout ReflectedLight reflectedLight) {
-					float dotNL = saturate(dot(geometry.normal, directLight.direction));
-					vec3 irradiance = dotNL * directLight.color;
-					reflectedLight.directDiffuse += irradiance * RECIPROCAL_PI * diffuseColor;
+				vec3 RE_Direct(const in IncidentLight directLight, const in vec3 normal) {
+					float dotNL = saturate(dot(normal, directLight.direction));
+
+					return dotNL * directLight.color * RECIPROCAL_PI * diffuse;
 				}
 
 				void main() {
-					vec4 diffuseColor = vec4(diffuse, opacity);
+					// If x is NaN, then y and z are also NaN.
+					vec3 normal = normalize(cross(dFdx(vViewPosition), dFdy(vViewPosition)));
 
-					ReflectedLight reflectedLight = ReflectedLight(vec3(0.0), vec3(0.0));
+					vec3 reflectedLight = vec3(0.0);
 
 					vec3 normal = normalize(cross(dFdx(vViewPosition), dFdy(vViewPosition)));
 
@@ -284,40 +277,34 @@ export default function ({ color, coords, opacity = 1 }, extent) {
 
 					IncidentLight directLight;
 
-					#if (NUM_POINT_LIGHTS > 0)
+					#if NUM_POINT_LIGHTS > 0
 						PointLight pointLight;
-						#pragma unroll_loop_start
 						for (int i = 0; i < NUM_POINT_LIGHTS; i++) {
 							pointLight = pointLights[i];
-							getPointLightInfo(pointLight, geometry, directLight);
-							RE_Direct(directLight, geometry, diffuseColor.rgb, reflectedLight);
+							getPointLightInfo(pointLight, directLight);
+							reflectedLight += RE_Direct(directLight, normal);
 						}
-						#pragma unroll_loop_end
 					#endif
-					#if (NUM_SPOT_LIGHTS > 0)
+					#if NUM_SPOT_LIGHTS > 0
 						SpotLight spotLight;
-						#pragma unroll_loop_start
 						for (int i = 0; i < NUM_SPOT_LIGHTS; i++) {
 							spotLight = spotLights[i];
-							getSpotLightInfo(spotLight, geometry, directLight);
-							RE_Direct(directLight, geometry, diffuseColor.rgb, reflectedLight);
+							getSpotLightInfo(spotLight, directLight);
+							reflectedLight += RE_Direct(directLight, normal);
 						}
-						#pragma unroll_loop_end
 					#endif
-					#if (NUM_DIR_LIGHTS > 0)
+					#if NUM_DIR_LIGHTS > 0
 						DirectionalLight directionalLight;
-						#pragma unroll_loop_start
 						for (int i = 0; i < NUM_DIR_LIGHTS; i++) {
 							directionalLight = directionalLights[i];
 							getDirectionalLightInfo(directionalLight, directLight);
-							RE_Direct(directLight, geometry, diffuseColor.rgb, reflectedLight);
+							reflectedLight += RE_Direct(directLight, normal);
 						}
-						#pragma unroll_loop_end
 					#endif
 
-					gl_FragColor = vec4(
-						reflectedLight.directDiffuse + reflectedLight.indirectDiffuse + ambientLightColor * RECIPROCAL_PI * diffuseColor.rgb,
-						diffuseColor.a
+					pc_fragColor = vec4(
+						reflectedLight + ambientLightColor * diffuse * RECIPROCAL_PI,
+						opacity
 					);
 				}
 			`