+++ CHANGES.md
@@ -8,6 +8,9 @@ Improvements:
 - Allow setting `maxSize` and `innerWidthMultiplier` from the translation layer. This was only possible to use with the "ES6 module" version of the library
 - Allow setting a custom extent (#52)
 
+Internals:
+- Convert all materials to ShaderMaterial/RawShaderMaterial (#53)
+
 1.0.3
 ---